refactor(server): move share types into package
