+++ packages/doc-server/package.json
@@ -10,6 +10,7 @@
     "./bridge": "./src/bridge.ts",
     "./collab": "./src/collab.ts",
     "./documents": "./src/documents.ts",
-    "./share": "./src/share.ts"
+    "./share": "./src/share.ts",
+    "./share-types": "./src/share-types.ts"
   }
 }
